driveToPosition() added to RobotHardware, with demo.
Function automates driving between the current 
position given by an instance of 
MecanumNavigation and a target position given 
by an instance of Navigation2D. 

Demo in Manual drives to (0,0,0) while the 
left and right bumpers and B are all held.

The key limitation to this routine is it solves the 
path between the current location and target 
while assuming that the Vx, Vy, and rotation rates are 
constant over the path.  This results in a single 
rotation around a point which restores the original 
position, which isn't an ideal path. The best case of 
this is when there is not rotation component, so 
only a translation occures (or a rotation around a 
infinitely distant point.)

Additional functions will be required to generate 
paths which are more linear or otherwise optimal.

Made tweak to Claw states so the open button 
always works after leaving claw testing mode.