Added waypoint driving logic, mecanumNavigation constructor/init
Added waypoint driving logic in state machine.
Converted angles to radians to be compatable with 
Navigation2D objects.
Currently, the waypoint logic is malfunctioning, 
and causing the robot to translate in wide 
circles once it drives to the first waypoint and 
rotates to face the second.

Added appropriate constructor, initialization, 
and update functions for mecanumNavigation and 
autoDrive.

Added documentation that theta is in radians CCW 
in MecanumNavigation.Navigation2D objects.