Manual mode autoDrive with dpad setup. Nav2D.toString() angle in degs
Origin can be reset by holding RB + LB and pressing 
either dpadUp or Y. This can be used to designate 
the center glyph column, while RB + LB and 
the dpadLeft and dpadRight are used to designate 
that the robot is lined up with the other two 
columns.
Alternate 'pickup position' can be set using RB + LB 
and pressing dpadDown.
Pressing dpadLeft or dpadRight will move to a 
position 6.5" to the left or right of the origin, for 
giving easy access to the 3 glyph columns.