Lower move tolerance 1>0.5 and smooth, tweak backup, correct claw open
Previously the backup waypoint wasn't working 
because the backup move distance was only 1 inch,
and the rotateThenDrive() tolerance was 1 inch, 
so no movement occured.
Since then, we've decreased our tolerance to 0.5 
and smoothed the proportional ramp down, 
and increased the backupDistance from 1 to 4 
inches.
Also added Current State telemetry output.