Removed old Auto opmodes
The old auto opmodes are now redundant. They 
scored 40 points, but the new opmode regularly 
scores 85 points.
The only potential reason to keep the old opmode 
would be to act as a fallback in case vuforia 
cannot be made to initialize, since it has 
intermittent problems in which the init() 
function times out.