Auto: zero nav on start(), show color telemetry in loop and init_loop
In Auto opmode, mecanum navigation is used and 
initialized in the init() function after the "Initialize" 
button is pressed.  Theoretically, if the wheels 
were turned after setting up the robot on the 
field, the initial nav position of the robot would 
be corrupted.  To deal with this, the mecanumNav 
is re-zeroed to the origin in start(), as soon as "Start" 
is selected.  This makes it possible to initialize the 
auto opmode on a cart before setting it up to ensure 
that it is functional and vuforia is ready.

Additionally, color telemetry data is now displayed 
both during the program loop and the init_loop, 
so diagnostics about intermittent color detection 
problems are easier.