BUGFIX: Eliminated infinite imu telemetry, added IMU update delay = 1.0
IMU telemetry can take a lot of time to generate.
BUGFIX: Was using a anonymous function in 
telemetry.addData, and was creating these in a loop, 
which caused more and more function calls per 
telemetry update, freezing the app. Very bad.

Also, added a delay to prevent updating the 
IMUUtilities object instance too often. Current min 
update delay is 1 sec, which seems huge, but still 
seems to cause a slight delay on the claw button 
controls.