Added Threaded Vufora Initialization to prevent hang in init().
One concern about this is the possibility of 
quickly hitting start to run the opmode without 
having time to initialize the vuforia object, 
inducing a null pointer Exception.
Steps have been taken to avoid this, so if the 
opmode gets to loop() while vuforia == null, 
errors should be caught and loading messages 
will be displayed.