Sped arm, but should descend faster, removed excess telemetry
Upward arm speed quite good, but downward 
could be increased a tad more.

An occasion was observed where the offset was 
not enough to clear a block. This could have had to 
do with the encoder initialization, but I suspect it 
has to do with the arm length used in calculations.
Either way, it may be prudent to slightly increase 
the size of the blocks to deal with those further 
out of spec.