Tweak ArmLength, MoveToPosition is default, sped descent.
System is working pretty well, but on the previous 
iteration, Level 4, which should be 18" high, 
was 18.75, and Level 4 offset, which should be 
19" high, extended higher and ran into the hard 
stop.  The key problem with this is the shaking 
induced at the top on an otherwise smooth system.
A mechanical solution would be to shave the 
hardstop down further to give more tolerance, but 
since I believe there is already enough headroom, 
the other alternative would be a software solution.
Decreasing ARM_TOP_TICKS will uniformly reduce 
angular motion by reducing ticksToDegrees.
Also, increasing ARM_LENGTH_INCHES will 
decrease motion such that heights will be uniformly 
reduced.  This may be the better option, since 
there seems to be 0.25" of error picked up every 
6" of travel.